+++ apps/portals/eliteportal/src/config/routeConstants.ts
@@ -1,0 +1,2 @@
+export const STUDY_DETAILS_PAGE_DETAILS_TAB_PATH = 'StudyDetails'
+export const STUDY_DETAILS_PAGE_DATA_TAB_PATH = 'StudyData'
